refactor[backend]: colocando o prefixo "api" nas rotas dos controllers e movendo SecurityConfig para o pacote de configuração
